Fix compatibility issue in 2021.2.4
